pyproject.toml
[project]
name = "polyedge"
version = "0.1.0"
description = "PolyEdge"
requires-python = ">=3.13"
dependencies = [
  "py-clob-client>=0.32.0",
  "python-dotenv>=1.2.0",
  "websockets>=15.0.1",
  "requests>=2.32.0",
  "web3>=7.15.0",
]

[dependency-groups]
dev = ["pre-commit>=4.5.0", "ruff>=0.15.0"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
